requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
Telethon==1.34.0
urllib3==2.2.1
